+++ src/agent_server/db/models/knowlege_file_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import JSON, Boolean, Column, DateTime, Float, Integer, String, func
+from sqlalchemy import JSON, Boolean, Column, DateTime, Float, Integer, BigInteger, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseEntity
@@ -13,7 +13,7 @@ class KnowledgeFile(BaseEntity, DateTimeMixin):
     __tablename__ = "knowledge_file"
     file_name: Mapped[str] = mapped_column(String(255), comment="文件名")
     file_ext: Mapped[str] = mapped_column(String(10), comment="文件扩展名")
-    kb_id: Mapped[int] = mapped_column(Integer, index=True, comment="所属知识库ID")
+    kb_id: Mapped[int] = mapped_column(BigInteger, index=True, comment="所属知识库ID")
     kb_name: Mapped[str] = mapped_column(String(50), comment="所属知识库名称")
     document_loader_name: Mapped[str] = mapped_column(String(50), comment="文档加载器名称")
     text_splitter_name: Mapped[str] = mapped_column(String(50), comment="文本分割器名称")
@@ -36,9 +36,9 @@ class FileDoc(BaseEntity):
     """
 
     __tablename__ = "file_doc"
-    kb_id: Mapped[int] = mapped_column(Integer, index=True, comment="知识库ID")
+    kb_id: Mapped[int] = mapped_column(BigInteger, index=True, comment="知识库ID")
     kb_name: Mapped[str] = mapped_column(String(50), comment="知识库名称")
-    file_id: Mapped[int] = mapped_column(Integer, index=True, comment="文件ID")
+    file_id: Mapped[int] = mapped_column(BigInteger, index=True, comment="文件ID")
     file_name: Mapped[str] = mapped_column(String(255), comment="文件名称")
     doc_id: Mapped[str] = mapped_column(String(50), comment="向量库文档ID")
     meta_data: Mapped[dict] = mapped_column(JSON, default={})
